requirements.txt
alabaster==0.7.12
argcomplete==2.0.0
asn1crypto==1.5.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.4
certifi==2022.9.24
chardet==5.1.0
click==8.1.6
cmake==3.26.3
colorama==0.4.6
command-not-found==0.3
crossref==0.1.2
crossrefapi==1.0.3
cryptography==38.0.4
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.8.0
distro-info==1.5+ubuntu0.23.10.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
duplicity==1.2.2
exceptiongroup==1.1.2
fasteners==0.17.3
filelock==3.12.0
fsspec==2023.4.0
future==0.18.2
GDAL==3.7.1
ghp-import==2.1.0
git-filter-repo==2.38.0
html5lib==1.1
httplib2==0.20.4
huggingface-hub==0.14.1
hypothesis==6.82.4
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
interchange==2021.0.4
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.2.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lit==16.0.2
livereload==2.6.3
lockfile==0.12.2
louis==3.26.0
lunr==0.6.2
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mechanize==0.4.8
mergedeep==1.3.4
mkdocs==1.4.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
packaging==23.1
pandas==2.0.2
pansi==2020.7.3
paramiko==2.12.0
pexpect==4.8.0
Pillow==10.0.0
pipreqs==0.4.13
pipx==1.2.0
platformdirs==3.5.1
pluggy==1.2.0
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
py2neo==2021.2.3
pyarrow==16.1.0
pycairo==1.24.0
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.46.0
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyrsistent==0.18.1
pytest==7.4.0
python-apt==2.6.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.0
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2022.10.31
requests==2.14.2
rich==13.3.1
roman==3.3
scour==0.38.2
SecretStorage==3.3.3
simplejson==3.19.1
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==5.3.0
sympy==1.11.1
systemd-python==235
texttable==1.6.7
timm==0.6.13
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.16
userpath==1.9.0
virtualenv==20.23.0
wadllib==1.3.6
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xmltodict==0.13.0
yarg==0.1.9
zipp==1.0.0
